backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
Pillow==10.0.1
stripe==6.7.0
requests==2.31.0
django-allauth==0.57.0
dj-rest-auth==5.0.2
cryptography==41.0.7
python-decouple==3.8
whitenoise==6.6.0 